requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
alembic==1.17.2
    # via mlflow
altair==6.0.0
    # via streamlit
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via starlette
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.14.3
    # via stock-curator (pyproject.toml)
blinker==1.9.0
    # via
    #   flask
    #   streamlit
cachetools==6.2.4
    # via
    #   google-auth
    #   mlflow-skinny
    #   mlflow-tracing
    #   streamlit
certifi==2025.11.12
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   flask
    #   mlflow-skinny
    #   streamlit
    #   uvicorn
cloudpickle==3.1.2
    # via mlflow-skinny
contourpy==1.3.2
    # via matplotlib
coverage==7.13.0
    # via pytest-cov
cryptography==46.0.3
    # via mlflow
cycler==0.12.1
    # via matplotlib
databricks-sdk==0.76.0
    # via
    #   mlflow-skinny
    #   mlflow-tracing
docker==7.1.0
    # via mlflow
exceptiongroup==1.3.1
    # via
    #   anyio
    #   pytest
fastapi==0.127.0
    # via mlflow-skinny
flask==3.1.2
    # via
    #   flask-cors
    #   mlflow
flask-cors==6.0.2
    # via mlflow
fonttools==4.61.1
    # via matplotlib
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via
    #   mlflow-skinny
    #   streamlit
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.28.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.187.0
    # via google-generativeai
google-auth==2.45.0
    # via
    #   databricks-sdk
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.3.0
    # via google-api-python-client
google-generativeai==0.8.6
    # via stock-curator (pyproject.toml)
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
graphene==3.4.3
    # via mlflow
graphql-core==3.2.7
    # via
    #   graphene
    #   graphql-relay
graphql-relay==3.2.0
    # via graphene
greenlet==3.3.0
    # via sqlalchemy
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
gunicorn==23.0.0
    # via mlflow
h11==0.16.0
    # via uvicorn
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
huey==2.5.5
    # via mlflow
idna==3.11
    # via
    #   anyio
    #   requests
importlib-metadata==8.7.1
    # via
    #   mlflow-skinny
    #   opentelemetry-api
iniconfig==2.3.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   altair
    #   flask
    #   pydeck
joblib==1.5.3
    # via scikit-learn
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
kiwisolver==1.4.9
    # via matplotlib
lxml==6.0.2
    # via lxml-html-clean
lxml-html-clean==0.4.3
    # via stock-curator (pyproject.toml)
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   mako
    #   werkzeug
matplotlib==3.10.8
    # via
    #   stock-curator (pyproject.toml)
    #   mlflow
mlflow==3.8.0
    # via stock-curator (pyproject.toml)
mlflow-skinny==3.8.0
    # via mlflow
mlflow-tracing==3.8.0
    # via mlflow
narwhals==2.14.0
    # via
    #   altair
    #   plotly
numpy==2.2.6
    # via
    #   stock-curator (pyproject.toml)
    #   contourpy
    #   matplotlib
    #   mlflow
    #   pandas
    #   pydeck
    #   scikit-learn
    #   scipy
    #   streamlit
    #   xgboost
nvidia-nccl-cu12==2.27.5
    # via xgboost
opentelemetry-api==1.39.1
    # via
    #   mlflow-skinny
    #   mlflow-tracing
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-proto==1.39.1
    # via
    #   mlflow-skinny
    #   mlflow-tracing
opentelemetry-sdk==1.39.1
    # via
    #   mlflow-skinny
    #   mlflow-tracing
opentelemetry-semantic-conventions==0.60b1
    # via opentelemetry-sdk
packaging==25.0
    # via
    #   altair
    #   gunicorn
    #   matplotlib
    #   mlflow-skinny
    #   mlflow-tracing
    #   plotly
    #   pytest
    #   streamlit
pandas==2.3.3
    # via
    #   stock-curator (pyproject.toml)
    #   mlflow
    #   streamlit
pillow==12.0.0
    # via
    #   matplotlib
    #   streamlit
plotly==6.5.0
    # via stock-curator (pyproject.toml)
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
proto-plus==1.27.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   databricks-sdk
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   mlflow-skinny
    #   mlflow-tracing
    #   opentelemetry-proto
    #   proto-plus
    #   streamlit
pyarrow==22.0.0
    # via
    #   mlflow
    #   streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   fastapi
    #   google-generativeai
    #   mlflow-skinny
    #   mlflow-tracing
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via pytest
pyparsing==3.3.0
    # via
    #   httplib2
    #   matplotlib
pytest==9.0.2
    # via
    #   stock-curator (pyproject.toml)
    #   pytest-cov
pytest-cov==7.0.0
    # via stock-curator (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   graphene
    #   matplotlib
    #   pandas
python-dotenv==1.2.1
    # via
    #   stock-curator (pyproject.toml)
    #   mlflow-skinny
pytz==2025.2
    # via
    #   stock-curator (pyproject.toml)
    #   pandas
pyyaml==6.0.3
    # via mlflow-skinny
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   stock-curator (pyproject.toml)
    #   databricks-sdk
    #   docker
    #   google-api-core
    #   mlflow-skinny
    #   streamlit
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
ruff==0.14.10
    # via stock-curator (pyproject.toml)
scikit-learn==1.7.2
    # via
    #   stock-curator (pyproject.toml)
    #   mlflow
scipy==1.15.3
    # via
    #   mlflow
    #   scikit-learn
    #   xgboost
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
soupsieve==2.8.1
    # via beautifulsoup4
sqlalchemy==2.0.45
    # via
    #   alembic
    #   mlflow
sqlparse==0.5.5
    # via mlflow-skinny
starlette==0.50.0
    # via fastapi
streamlit==1.52.2
    # via stock-curator (pyproject.toml)
tenacity==9.1.2
    # via streamlit
threadpoolctl==3.6.0
    # via scikit-learn
toml==0.10.2
    # via streamlit
tomli==2.3.0
    # via
    #   alembic
    #   coverage
    #   pytest
tornado==6.5.4
    # via streamlit
tqdm==4.67.1
    # via google-generativeai
typing-extensions==4.15.0
    # via
    #   alembic
    #   altair
    #   anyio
    #   beautifulsoup4
    #   cryptography
    #   exceptiongroup
    #   fastapi
    #   google-generativeai
    #   graphene
    #   grpcio
    #   mlflow-skinny
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   starlette
    #   streamlit
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.2
    # via
    #   docker
    #   requests
uvicorn==0.40.0
    # via mlflow-skinny
watchdog==6.0.0
    # via streamlit
werkzeug==3.1.4
    # via
    #   flask
    #   flask-cors
xgboost==3.1.2
    # via stock-curator (pyproject.toml)
zipp==3.23.0
    # via importlib-metadata
